+++ api/api/controllers/cluster_controller.py
@@ -684,7 +684,7 @@ def get_conf_validation_node(node_id, pretty=False, wait_for_complete=False):
 
 
 @exception_handler
-def get_node_config(node_id, component, configuration, wait_for_complete=False):
+def get_node_config(node_id, component, configuration, wait_for_complete=False, pretty=False):
     """Get active configuration in node node_id [on demand]
 
     Returns the requested configuration.
@@ -692,21 +692,19 @@ def get_node_config(node_id, component, configuration, wait_for_complete=False):
     :param wait_for_complete: Disable timeout response
     :param node_id: Cluster node name.
     :param component: Specified component.
-    :param component: Specified configuration.
+    :param configuration: Specified configuration.
     """
-
-    # WIP
-
-    f_kwargs = {'agent_id': node_id,
-                'component': component,
-                'configuration': configuration
+    f_kwargs = {'component': component,
+                'configuration': configuration,
+                'agent_id': node_id
                 }
 
-    dapi = DistributedAPI(f=cluster.get_config_ondemand,
+    dapi = DistributedAPI(f=manager.get_config,
                           f_kwargs=remove_nones_to_dict(f_kwargs),
                           request_type='distributed_master',
                           is_async=False,
                           wait_for_complete=wait_for_complete,
+                          pretty=pretty,
                           logger=logger
                           )
     data = loop.run_until_complete(dapi.distribute_function())