+++ framework/wazuh/manager.py
@@ -21,7 +21,7 @@ from typing import Dict
 import fcntl
 
 from wazuh import common
-from wazuh.exception import WazuhException, WazuhError, WazuhInternalError
+from wazuh.exception import WazuhException
 from wazuh.utils import previous_month, cut_array, sort_array, search_array, tail, load_wazuh_xml
 from wazuh import configuration
 
@@ -183,29 +183,44 @@ def ossec_log_summary(months=3):
     return categories
 
 
-def upload_file(path='', content='', overwrite=False):
+def upload_file(tmp_file, path, content_type, overwrite=False):
     """
     Updates a group file
 
+    :param tmp_file: Relative path of file name from origin
     :param path: Path of destination of the new file
-    :param content: Content of the file to be uploaded
+    :param content_type: Content type of file from origin
     :param overwrite: True for updating existing files, False otherwise
     :return: Confirmation message in string
     """
-    if not path:
-        raise WazuhError(1908)
+    try:
+        # if file already exists and overwrite is False, raise exception
+        if not overwrite and exists(join(common.ossec_path, path)):
+            raise WazuhException(1905)
 
-    if not content:
-        raise WazuhException(1909)
+        try:
+            with open(join(common.ossec_path, tmp_file)) as f:
+                file_data = f.read()
+        except IOError:
+            raise WazuhException(1005)
+        except Exception:
+            raise WazuhException(1000)
 
-    # if file already exists and overwrite is False, raise exception
-    if not overwrite and exists(join(common.ossec_path, path)):
-        raise WazuhError(1905)
-    
-    if path.split('/')[1] == 'rules':
-        return upload_list(content, path)
-    else:
-        return upload_xml(content, path)
+        if len(file_data) == 0:
+            raise WazuhException(1112)
+
+        if content_type == 'application/xml':
+            return upload_xml(file_data, path)
+        elif content_type == 'application/octet-stream':
+            return upload_list(file_data, path)
+        else:
+            raise WazuhException(1016)
+    finally:
+        # delete temporary file from API
+        try:
+            remove(join(common.ossec_path, tmp_file))
+        except OSError:
+            raise WazuhException(1903)
 
 
 def upload_xml(xml_file, path):
@@ -235,27 +250,27 @@ def upload_xml(xml_file, path):
             tmp_file.write(final_xml)
         chmod(tmp_file_path, 0o640)
     except IOError:
-        raise WazuhInternalError(1005)
+        raise WazuhException(1005)
     except ExpatError:
-        raise WazuhError(1113)
+        raise WazuhException(1113)
     except Exception as e:
-        raise WazuhInternalError(1000, str(e))
+        raise WazuhException(1000, str(e))
 
     try:
         # check xml format
         try:
             load_wazuh_xml(tmp_file_path)
         except Exception as e:
-            raise WazuhError(1113, str(e))
+            raise WazuhException(1113, str(e))
 
         # move temporary file to group folder
         try:
             new_conf_path = join(common.ossec_path, path)
             move(tmp_file_path, new_conf_path)
         except Error:
-            raise WazuhInternalError(1016)
+            raise WazuhException(1016)
         except Exception:
-            raise WazuhInternalError(1000)
+            raise WazuhException(1000)
 
         return 'File updated successfully'
 
@@ -323,9 +338,7 @@ def get_file(path, validation=False):
         with open(full_path) as f:
             output = f.read()
     except IOError:
-        raise WazuhInternalError(1005)
-    except Exception:
-        raise WazuhInternalError(1000)
+        raise WazuhException(1005)
 
     return output
 
@@ -383,9 +396,9 @@ def delete_file(path):
         try:
             remove(full_path)
         except IOError:
-            raise WazuhInternalError(1907)
+            raise WazuhException(1907)
     else:
-        raise WazuhError(1906)
+        raise WazuhException(1906)
 
     return 'File was deleted'
 
@@ -396,27 +409,6 @@ def restart():
 
     :return: Confirmation message.
     """
-<<<<<<< HEAD
-    # execq socket path
-    socket_path = common.EXECQ
-    # msg for restarting Wazuh manager
-    msg = 'restart-wazuh '
-    # initialize socket
-    if exists(socket_path):
-        try:
-            conn = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM)
-            conn.connect(socket_path)
-        except socket.error:
-            raise WazuhInternalError(1902)
-    else:
-        raise WazuhInternalError(1901)
-
-    try:
-        conn.send(msg.encode())
-        conn.close()
-    except socket.error:
-        raise WazuhInternalError(1014)
-=======
     lock_file = open(execq_lockfile, 'a+')
     fcntl.lockf(lock_file, fcntl.LOCK_EX)
     try:
@@ -442,7 +434,6 @@ def restart():
     finally:
         fcntl.lockf(lock_file, fcntl.LOCK_UN)
         lock_file.close()
->>>>>>> dev-flask-poc
 
     return "Restarting manager"
 
@@ -481,24 +472,6 @@ def validation():
     lock_file = open(execq_lockfile, 'a+')
     fcntl.lockf(lock_file, fcntl.LOCK_EX)
     try:
-<<<<<<< HEAD
-        remove(api_socket_path)
-    except OSError:
-        if exists(api_socket_path):
-            raise WazuhInternalError(1014)
-
-    # up API socket
-    try:
-        api_socket = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM)
-        api_socket.bind(api_socket_path)
-        # timeout
-        api_socket.settimeout(5)
-    except socket.error:
-        raise WazuhInternalException(1013)
-
-    # connect to execq socket
-    if exists(execq_socket_path):
-=======
         # sockets path
         api_socket_path = join(common.ossec_path, 'queue/alerts/execa')
         execq_socket_path = common.EXECQ
@@ -506,7 +479,6 @@ def validation():
         execq_msg = 'check-manager-configuration '
 
         # remove api_socket if exists
->>>>>>> dev-flask-poc
         try:
             remove(api_socket_path)
         except OSError as e:
@@ -520,39 +492,6 @@ def validation():
             # timeout
             api_socket.settimeout(5)
         except socket.error:
-<<<<<<< HEAD
-            raise WazuhInternalError(1013)
-    else:
-        raise WazuhInternalError(1901)
-
-    # send msg to execq socket
-    try:
-        execq_socket.send(execq_msg.encode())
-        execq_socket.close()
-    except socket.error:
-        raise WazuhInternalError(1014)
-    finally:
-        execq_socket.close()
-
-    # if api_socket receives a message, configuration is OK
-    try:
-        buffer = bytearray()
-        # receive data
-        datagram = api_socket.recv(4096)
-        buffer.extend(datagram)
-    except socket.timeout:
-        raise WazuhInternalError(1014)
-    finally:
-        api_socket.close()
-        # remove api_socket
-        if exists(api_socket_path):
-            remove(api_socket_path)
-
-    try:
-        response = _parse_execd_output(buffer.decode('utf-8').rstrip('\0'))
-    except (KeyError, json.decoder.JSONDecodeError):
-        raise WazuhInternalError(1904)
-=======
             raise WazuhException(1013)
 
         # connect to execq socket
@@ -595,7 +534,6 @@ def validation():
     finally:
         fcntl.lockf(lock_file, fcntl.LOCK_UN)
         lock_file.close()
->>>>>>> dev-flask-poc
 
     return response
 
@@ -623,8 +561,8 @@ def _parse_execd_output(output: str) -> Dict:
     return response
 
 
-def get_config(component, config):
+def get_config(component, config, agent_id='000'):
     """
     Returns active configuration loaded in manager
     """
-    return configuration.get_active_configuration(agent_id='000', component=component, configuration=config)
+    return configuration.get_active_configuration(agent_id=agent_id, component=component, configuration=config)