+++ api/api/spec/changes.md
@@ -133,6 +133,9 @@ a human readable message is shown, the new field `message` will be used instead.
 ### GET /cluster/{node_id}/configuration/validation
 * Now errors are shown in a different schema with a HTTP status 400. See spec for more details.
 
+### GET /cluster/{node_id}/configuration/{component}/{configuration}
+* Now errors are shown in a different schema with a HTTP status 400. See spec for more details.
+
 
 ## Experimental
 ### General